*** empty log message ***

+++ org.eclipse.cdt.old/org.eclipse.cdt.cpp.miners/org/eclipse/cdt/cpp/miners/managedproject/ManagedProjectMiner.java
@@ -2,6 +2,7 @@ package com.ibm.cpp.miners.managedproject;
 
 import com.ibm.dstore.core.model.*;
 import com.ibm.dstore.core.miners.miner.*;
+import com.ibm.cpp.miners.managedproject.amparser.*;
 
 import java.util.*;
 import java.io.*;
@@ -15,16 +16,16 @@ public class ManagedProjectMiner extends Miner
  
  public void extendSchema(DataElement schemaRoot)
  {
-  DataElement projectD = _dataStore.find(schemaRoot, DE.A_NAME, "CProject");
+  DataElement projectD = _dataStore.find(schemaRoot, DE.A_NAME, "Project");
   createCommandDescriptor(projectD, "Manage Project", "C_MANAGE_PROJECT");
-  
     
-  
-  DataElement managedProjectD  = _dataStore.createObject(schemaRoot, DE.T_OBJECT_DESCRIPTOR, "Managed Project");
+  DataElement managedProjectD  = _dataStore.createObject(schemaRoot, DE.T_OBJECT_DESCRIPTOR, Am.MANAGED_PROJECT);
   createCommandDescriptor(managedProjectD, "Unmanage Project", "C_UNMANAGE_PROJECT");
-  DataElement targetD          = _dataStore.createObject(schemaRoot, DE.T_OBJECT_DESCRIPTOR, "Project Target");
-  DataElement targetAttributeD = _dataStore.createObject(schemaRoot, DE.T_OBJECT_DESCRIPTOR, "Target Attribute");
-  DataElement targetOptionD    = _dataStore.createObject(schemaRoot, DE.T_OBJECT_DESCRIPTOR, "Target Option");
+  DataElement targetD          = _dataStore.createObject(schemaRoot, DE.T_OBJECT_DESCRIPTOR, Am.PROJECT_TARGET);
+  
+  DataElement targetAttributeTypeD = _dataStore.createObject(schemaRoot, DE.T_OBJECT_DESCRIPTOR, Am.TARGET_ATTRIBUTE_TYPE);
+  DataElement targetAttributeD = _dataStore.createObject(schemaRoot, DE.T_OBJECT_DESCRIPTOR, Am.TARGET_ATTRIBUTE);
+  DataElement targetOptionD    = _dataStore.createObject(schemaRoot, DE.T_OBJECT_DESCRIPTOR, Am.TARGET_OPTION);
   
   DataElement workspaceD       = _dataStore.find(schemaRoot, DE.A_NAME, "Workspace", 1);
   DataElement projectFileD     = _dataStore.find(schemaRoot, DE.A_NAME, "Project File", 1);
@@ -33,9 +34,9 @@ public class ManagedProjectMiner extends Miner
   _dataStore.createReference(managedProjectsD, managedProjectD);
   _dataStore.createReference(managedProjectsD, targetD);
   _dataStore.createReference(managedProjectsD, projectFileD);
+  _dataStore.createReference(managedProjectsD, targetAttributeTypeD);
   _dataStore.createReference(managedProjectsD, targetAttributeD);
   _dataStore.createReference(managedProjectsD, targetOptionD);
-  
   
   _dataStore.createReference(workspaceD, managedProjectsD);
 
@@ -43,58 +44,43 @@ public class ManagedProjectMiner extends Miner
   createCommandDescriptor(targetD, "Build Target", "C_BUILD_TARGET");
   createCommandDescriptor(targetD, "Modify Target", "C_MODIFY_TARGET");
   createCommandDescriptor(targetD, "Remove Target", "C_REMOVE_TARGET");
+ 
  }
  
  public DataElement handleCommand(DataElement theCommand)
  {
-  String name          = getCommandName(theCommand);
-  DataElement  status  = getCommandStatus(theCommand);
+  String          name = getCommandName(theCommand);
+  DataElement   status = getCommandStatus(theCommand);
   DataElement  project = getCommandArgument(theCommand, 0);
- 
+  
+  if (!project.getType().equals("project"))
+   return status;
+  
   if (name.equals("C_UNMANAGE_PROJECT"))
-  {
    _dataStore.deleteObject(project.getParent(), project);
-  }
-  
   else if (name.equals("C_MANAGE_PROJECT"))
-  {
-   String fileLocation = _dataStore.getAttribute(DataStoreAttributes.A_PLUGIN_PATH) + "/com.ibm.cpp.miners/sample";
-
-   DataElement workspace = project.getParent();
-   DataElement managedProject = _dataStore.createObject(workspace, "Managed Project", "MyProject", fileLocation + File.separator + "Makefile.am");
-   DataElement hello          = _dataStore.createObject(managedProject, "Project Target", "hello");
-   DataElement goodbye        = _dataStore.createObject(managedProject, "Project Target", "goodbye");
+   parseAmFile(project);
+  else if (name.equals("C_REFRESH"))
+   refreshProject(project);
   
-   DataElement helloSources        = _dataStore.createObject(hello, "Target Attribute", "SOURCES");
-   DataElement helloLdadd          = _dataStore.createObject(hello, "Target Attribute", "LDADD");
-   DataElement helloLdFlags        = _dataStore.createObject(hello, "Target Attribute", "LDFLAGS");
-   DataElement helloDependencies   = _dataStore.createObject(hello, "Target Attribute", "DEPENDENCIES");
-   DataElement goodbyeSources      = _dataStore.createObject(goodbye, "Target Attribute", "SOURCES");
-
-   
-  _dataStore.createObject(helloSources, "Project File", "aux.cpp");
-  _dataStore.createObject(helloSources, "Project File", "aux.h");
-  _dataStore.createObject(helloSources, "Project File", "hello.cpp");
-  _dataStore.createObject(helloLdadd, "Project File", "util.a");
-  _dataStore.createObject(helloLdFlags, "Target Option", "-all-static");
-  _dataStore.createReference(helloDependencies, goodbye);
-  
-  _dataStore.createObject(goodbyeSources, "Project File", "bye.c");
-  _dataStore.createObject(goodbyeSources, "Project File", "bye.h");
-
-
-
-  DataElement utilManagedProject = _dataStore.createObject(managedProject, "Managed Project", "util", fileLocation + File.separator + "util" + File.separator + "Makefile.am");
-  DataElement util = _dataStore.createObject(utilManagedProject, "Project Target", "util");
-  DataElement utilSources        = _dataStore.createObject(util, "Target Attribute", "SOURCES");
-  DataElement utilLibadd         = _dataStore.createObject(util, "Target Attribute", "LIBADD");
-  _dataStore.createObject(utilSources, "Project File", "util.c");
-  _dataStore.createObject(utilSources, "Project File", "util.h");
-  _dataStore.createObject(utilLibadd, "Project File", "fooX.o");
- }
-      
   status.setAttribute(DE.A_NAME, getLocalizedString("model.done"));
   return status;
  }
+
+ private DataElement parseAmFile(DataElement theUnmanagedProject)
+ {
+  AmParser theParser = new AmParser(theUnmanagedProject);
+  DataElement theManagedProject = theParser.parse();
+  _dataStore.refresh(theManagedProject);
+  _dataStore.refresh(theManagedProject.getParent());
+  return theManagedProject;
+ }
+
+ private DataElement refreshProject(DataElement theUnmanagedProject)
+ {
+  System.out.println("Got it");
+  return null;
+ }
+ 
 }
